Rename buildrun.sh to run.sh
I feel like the purpose of buildrun.sh is to be the 'dumb' option, i.e.
accessible to people who have no interest in learning how to use Docker
or the many finnicky options that our image requires to be run. I think
generally such a script would be called 'run.sh' so that's what I'm
renaming it to.