Change to Use Host-Linked Dir
This appears to allow the container to use the full storage capacity of
the host... we hope. At least, that's what the results of the Wikibrain
Loader diagnostics show. It's kind of a pain to get it to work, since
the conatiner-side folder is wiped at build time, therefore we have to
copy over everything at run-time.